Move plugin-ast fixture/test to individual locations
