--- packages/knip/fixtures/plugin-ast/next.config.ts
@@ -1,7 +1,0 @@
-import 'next';
-
-const nextConfig = {
-  pageExtensions: ['page.tsx'],
-};
-
-export default nextConfig;
