--- packages/knip/fixtures/plugin-ast/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@fixtures/plugin-ast",
-  "dependencies": {
-    "@tanstack/react-router": "*",
-    "@tanstack/router-plugin": "*",
-    "@vitejs/plugin-react": "*",
-    "astro": "*",
-    "next": "*",
-    "sst": "*",
-    "vite": "*"
-  }
-}
