--- packages/knip/fixtures/plugin-ast/routes/index.ts
@@ -1,1 +1,0 @@
-import '@tanstack/react-router';
