+++ packages/knip/fixtures/plugins/sst/handlers/auth.handler.ts
