+++ packages/knip/fixtures/plugins/sst/handlers/auth.ts
