+++ packages/knip/fixtures/plugins/sst/handlers/some-route-handler.ts
