+++ packages/knip/fixtures/plugins/sst/package.json
@@ -1,4 +1,7 @@
 {
   "name": "@fixtures/sst",
-  "version": "*"
+  "version": "*",
+  "dependencies": {
+    "sst": "*"
+  }
 }
