+++ packages/knip/fixtures/plugins/sst/stacks/AuthHandlerStack.ts
