+++ packages/knip/fixtures/plugins/sst/stacks/AuthStack.ts
