+++ packages/knip/fixtures/plugins/starlight/astro.config.mjs
