+++ packages/knip/fixtures/plugins/starlight/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@fixtures/starlight",
+  "version": "*",
+  "dependencies": {
+    "@astrojs/starlight": "*",
+    "astro": "*"
+  }
+}
