+++ packages/knip/fixtures/plugins/tanstack-router/package.json
@@ -2,6 +2,9 @@
   "name": "@fixtures/tanstack-router",
   "version": "*",
   "dependencies": {
-    "@tanstack/react-router": "*"
+    "@tanstack/react-router": "*",
+    "@tanstack/router-plugin": "*",
+    "@vitejs/plugin-react": "*",
+    "vite": "*"
   }
 }
