+++ packages/knip/fixtures/plugins/tanstack-router/vite.config.ts
