+++ packages/knip/src/plugins/starlight/resolveFromAST.ts
